requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.9.2
aiofiles=23.2.1=pypi_0
annotated-doc=0.0.4=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
bidict=0.23.1=pypi_0
blas=1.0=mkl
brotlicffi=1.0.9.2=py39h885b0b7_2
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.12.2=haa95532_0
certifi=2025.10.5=py39haa95532_0
cffi=2.0.0=py39h02ab6af_0
charset-normalizer=3.4.4=py39haa95532_0
click=8.1.8=pypi_0
controlnet-aux=0.0.3=pypi_0
cuda-cccl=12.9.27=0
cuda-cccl_win-64=12.9.27=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.9.79=0
cuda-runtime=11.8.0=0
cuda-version=12.9=3
diffusers=0.16.1=pypi_0
einops=0.8.1=pypi_0
expat=2.7.3=h885b0b7_4
fastapi=0.124.2=pypi_0
ffmpy=1.0.0=pypi_0
filelock=3.17.0=py39haa95532_0
flask=2.2.3=pypi_0
flask-socketio=5.5.1=pypi_0
flaskwebgui=0.3.5=pypi_0
freetype=2.13.3=h0620614_0
fsspec=2025.10.0=pypi_0
giflib=5.2.2=h7edc060_0
gmp=6.3.0=h537511b_0
gmpy2=2.2.1=py39h827c3e9_0
gradio=4.44.1=pypi_0
gradio-client=1.3.0=pypi_0
huggingface-hub=0.19.4=pypi_0
idna=3.11=py39haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.6=py39haa95532_0
jpeg=9f=ha349fce_0
lama-cleaner=1.2.5=pypi_0
lcms2=2.16=hb4a4139_0
lerc=3.0=hd77b12b_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.10.19=0
libcurand-dev=10.3.10.19=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_1
libexpat=2.7.3=h885b0b7_4
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.50=h46444df_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
libzlib=1.3.1=h02ab6af_0
loguru=0.7.3=pypi_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.5=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h827c3e9_2
mkl_fft=1.3.11=py39h827c3e9_0
mkl_random=1.2.8=py39hc64d2fc_0
mpc=1.3.1=h827c3e9_0
mpfr=4.2.1=h56c3642_0
mpmath=1.3.0=py39haa95532_0
networkx=3.2.1=py39haa95532_0
numpy=2.0.1=py39h055cbcc_1
numpy-base=2.0.1=py39h65a83cf_1
omegaconf=2.3.0=pypi_0
opencv-python=4.12.0.88=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.18=h543e019_0
orjson=3.11.5=pypi_0
pandas=2.3.3=pypi_0
piexif=1.1.3=pypi_0
pillow=11.1.0=py39h096bfcc_0
pip=25.3=pyhc872135_0
pycparser=2.23=py39haa95532_0
pydantic=2.12.5=pypi_0
pydantic-core=2.41.5=pypi_0
pydub=0.25.1=pypi_0
pysocks=1.7.1=py39haa95532_1
python=3.9.25=h716150d_1
python-engineio=4.12.3=pypi_0
python-multipart=0.0.20=pypi_0
python-socketio=5.15.0=pypi_0
pytorch=2.5.1=py3.9_cuda11.8_cudnn9_0
pytorch-cuda=11.8=h24eeafa_6
pytorch-mutex=1.0=cuda
pyyaml=6.0.2=py39h827c3e9_0
regex=2025.11.3=pypi_0
requests=2.32.5=py39haa95532_0
ruff=0.14.9=pypi_0
safetensors=0.7.0=pypi_0
segment-anything=1.0=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=80.9.0=py39haa95532_0
shellingham=1.5.4=pypi_0
simple-websocket=1.1.0=pypi_0
sqlite=3.51.0=hda9a48d_0
starlette=0.49.3=pypi_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=h59b6b97_0
timm=1.0.22=pypi_0
tk=8.6.15=hf199647_0
tokenizers=0.13.3=pypi_0
tomlkit=0.12.0=pypi_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
transformers=4.27.4=pypi_0
typer=0.20.0=pypi_0
typing-inspection=0.4.2=pypi_0
typing_extensions=4.15.0=py39haa95532_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.6.2=pypi_0
uvicorn=0.38.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
websockets=12.0=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.45.1=py39haa95532_0
whichcraft=0.6.1=pypi_0
win32-setctime=1.2.0=pypi_0
win_inet_pton=1.1.0=py39haa95532_1
xz=5.6.4=h4754444_1
yacs=0.1.8=pypi_0
yaml=0.2.5=he774522_0
zlib=1.3.1=h02ab6af_0
zstd=1.5.7=h56299aa_0
